Fix item image upload miss delete problem Caused by deleting non jpg image after converting to jpg, it will delete the image file if it didnt perform any format conversion

+++ assets/classes/ImageFileHandler.class.php
@@ -165,7 +165,7 @@ class ImageFileHandler{
         $fullPath = $path . "/" . $fileName . "." . $extension;
         $binary = imagecreatefromstring(file_get_contents($fullPath));
         imageJpeg($binary, $path . $fileName . '.jpg', $imageQuality);
-        unlink($fullPath);
+        if($extension != "jpg") unlink($fullPath);
     }
 
     private function optimizeImageJPG($path, $fileName){ // Only accept jpg format // Old version backup
